setup for confusion matrix
